change texture W/H to uint8_t

+++ include/hardware/gpu.h
@@ -20,7 +20,7 @@ typedef struct{
 
 typedef struct {
     uint8_t u, v;
-    uint16_t w, h;
+    uint8_t w, h;
     uint16_t page, clut; // Page and clut values for rendering
 } TextureInfo;
